+++ src/app/forms/checkbox-page/checkbox-page.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
+import { ReactiveFormsModule } from '@angular/forms';
 
 import { CheckboxModule, SectionModule } from '@swimlane/ngx-ui';
 import { PrismModule } from '../../common/prism/prism.module';
@@ -10,6 +11,14 @@ import { CheckboxPageComponent } from './checkbox-page.component';
 
 @NgModule({
   declarations: [CheckboxPageComponent],
-  imports: [CommonModule, FormsModule, PrismModule, SectionModule, CheckboxModule, CheckboxPageRoutingModule]
+  imports: [
+    CommonModule,
+    FormsModule,
+    PrismModule,
+    SectionModule,
+    CheckboxModule,
+    CheckboxPageRoutingModule,
+    ReactiveFormsModule
+  ]
 })
 export class CheckboxPageModule {}