+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## HEAD (unreleased)
 
+- Feature: added `ngx-time` component
 - Enhancement: Improved keyboard accessibility for `ngx-select`
 - Enhancement: Add ability to unlock `ngx-input` for passwords
+- Enhancement: Add `checked` and `checkedChange` to `ngx-checkbox`
+- Enhancement: Keyboard accessibility for `ngx-checkbox`
 
 ## 36.1.0 (2021-12-10)